+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/remainingnumber/specialleave/empinfo/grantremainingdata/SpecialLeaveGrantRemainingData.java
@@ -39,8 +39,8 @@ public class SpecialLeaveGrantRemainingData extends AggregateRoot {
 
 	public static SpecialLeaveGrantRemainingData createFromJavaType(String specialId,String cid,String employeeId, int specialLeaveCode,
 			GeneralDate grantDate, GeneralDate deadlineDate, int expirationStatus, int registerType,
-			int dayNumberOfGrant, Integer timeOfGrant, Double dayNumberOfUse, Integer timeOfUse,
-			Double useSavingDays, int numberOverdays, Integer timeOver, Double dayNumberOfRemain,
+			double dayNumberOfGrant, Integer timeOfGrant, double dayNumberOfUse, Integer timeOfUse,
+			Double useSavingDays, double numberOverdays, Integer timeOver, double dayNumberOfRemain,
 			Integer timeOfRemain) {
 		SpecialLeaveGrantRemainingData domain = new SpecialLeaveGrantRemainingData();
 		domain.specialId = specialId;
